requirements.txt
Flask>=2.3.2,<2.4
Werkzeug>=2.3.6,<2.4
gunicorn>=20.1.0,<21.0
pymongo>=4.3.3,<5.0
notion-client>=2.0.0,<3.0
google-auth>=2.17.3,<3.0
google-auth-oauthlib>=1.0.0,<2.0
google-auth-httplib2>=0.1.0,<0.2
google-api-python-client>=2.86.0,<3.0
flask-limiter>=3.3.1,<4.0
python-dotenv>=1.0.0,<2.0


